add cmdarg_parse_to_positional function
this function assumes that once you've seen one positional parameter, all
following parameters are also positional even if they look like a supported
script flag. In other words, it disallows interleving flags and positional
operands.  This is useful for parsing command line arguments which take another
command as an operand.

for example

./myscript --script-option=1 ./other-command --inner-script-flag

would be parsed as:

script-option = 1
cmdarg_argv = ./other-command --inner-script-flag

note that cmdarg_parse also supports passing this type of data but requires the
user to type -- in between the flags and the positional parameters as such:

./myscript --script-option=1 -- ./other-command --inner-script-flag